Fix reading from a pipe on freebsd
On freebsd, this currently fails:

cat foo.pcap | termshark -r -

When termshark is asked to read from stdin, it currently dupes stdin to
another descriptor X and rewrites the argument as /dev/fd/X. This was
done to avoid a conflict with the UI input on descriptor 0, though this
may not be necessary now (just open /dev/tty on another descriptor for
UI interaction). When tshark/dumpcap is started to read the pcap data
and to feed it to termshark, the argument /dev/fd/X is passed
through. But on freebsd, this doesn't work

cat foo.pcap | tshark -r /dev/fd/0

I'm not sure why. I can workaround this problem because the default
capture program on Unix OSes is for termshark to run itself in a special
mode - it will then simply pass through args to dumpcap, and then if
that fails, tshark. Since the code for this special mode is inside
termshark, I take the opportunity to rewrite /dev/fd/X to "-", and
dup2() descriptor X to 0 - so dumpcap is run like this instead

dumpcap -r -

dumpcap on freebsd understands "-" to mean stdin, and so this approach
works.