Save per-terminal-mode theme settings
At this point, termshark doesn't insist that a theme has to support
every typical terminal color mode - 16-colors, 256-colors,
truecolor. Most of the current themes are for 256-colors, and will also
work in truecolor mode. If the user has theme dracula set in 256-colors
mode, then runs termshark in a cut-down terminal on the same
machine (e.g. an xterm with 16 colors), then the theme won't
load. Rather than having that silently fail, I'm now saving a theme
setting per terminal mode, under toml keys theme-16, theme-256 and
theme-truecolor. When the theme is set, a message is displayed to make
it clear the theme is set for this color mode only. It's still not
completely obvious though - if the user set a theme in one terminal
emulator, the user might still wonder why it doesn't load up in the
cut-down terminal emulator. I'm not displaying a message - I don't want
to make the user click it away. Maybe it would help to have a more
obvious message buffer for short messages of interest to the user that
will disappear on their own...