Simplify file descriptor handling for piped input
I made this overly complicated because of a misunderstanding. Say the
user starts termshark like this

$ cat foo.pcap | termshark -i -

Termshark would understand "-" to mean stdin; would then dup()
descriptor 0 (say to X), and internally track the input source to
termshark as type:pipe name:/dev/fd/X. Then when dumpcap/tshark is
started to read the packets, its input is provided as /dev/fd/X, and the
descriptor X can be read because it's available in the child process.

I did this dup() step because I had problems just passing through
/dev/fd/0, and I assumed it was a clash with termshark's use of
descriptor 0 for the UI. But of course termshark isn't using descriptor
0 for the UI - it opens /dev/tty for read and write via tcell, and those
descriptors handle UI input and output. If termshark is invoked as
above, then descriptor 0 will be the read end of a pipe, anyway. What
was going wrong is that when a Go *Cmd is executed, unless you
explicitly set cmd.{Stdin,Stderr,Stdout}, they are set to
/dev/null. That meant when I ran dumpcap/tshark to feed packets from
/dev/fd/0 to termshark, descriptor 0 was useless. This wasn't a problem
for descriptor X > 2 because Go seems to have the child inherit these by
default.

The fix is to explicitly set cmd.{Stdin,Stderr,Stdout} to those from the
parent (Stdin is the one that really matters in this case)