More changes to the packet loaders
I'm still not satisfied. The crux is that if proc.Wait() is called on a
different goroutine to the one issuing the kill, it's possible that the
pid has gone - and in an extreme case, be recycled, when the kill is
issued. This change reorganizes the pcap/pdml loader so that a new
reaper goroutine will wait for the following information:

- wait on a channel for each of the pdml/pcap processes
- wait for the stage 2 context to be cancelled

A goroutine is run for handling the creation of the pdml and pcap
reading processes. If stage 2 is cancelled before they are started, then
the pdml/pcap channels are closed. The reaper goroutine will receive a
zero-value pid on these channels, and know the processes were not
started. If the pdml/pcap goroutines proceed before stage 2
cancellation, but something goes wrong starting these processes, the
reaper goroutine will again receive zero-value pids. If the processes
are started, then the reaper will receive their pids, and will know that
Wait() must be called later. On a stage 2 cancellation, which is now
guaranteed to happen at the latest by the end of the loading process,
the reaper will know if it needs to call Wait() on the pdml and pcap
processes.