Fix a regression I introduced with recent loader changes
By reordering some goroutines, I captured the value of the display
filter to use for the pcap/pdml load before it was set in the stage 1
load. The result was that if you loaded a pcap, set a display filter,
then navigated to a packet and selected stream reassembly, the wrong
stream would be chosen.