Better tab-completion behavior when multiple options match
Currently, the minibuffer will show possible completions for commands
when you hit the tab key. One is always in focus, and if you hit enter
then that choice will be selected and the minibuffer will be
populated. But if you hit tab, it's not intuitive to select the option
in focus if more than one candidate could be chosen. For example - say
you type

> cl

The possible completions are "clear" and "clear-filter". Hitting tab
should change the minibuffer so that it shows:

> clear

This change makes that happen.