Fix occasional failed execution of jump to global mark
Sometimes I found that jumping to a global mark in a different pcap to
the one currently loaded would fail. This happened because the global
mark code runs in a callback from the pcap loader, executed when the
load completes, and was run concurrently with other handlers including
the one that updated the packet list with the final batch of data from
the loader. If the global mark callback ran first and the jump position
was close to the end of the pcap, it might require the last batch of
packets to have been processed and added to various structures. The fix
is to run the handlers serially - I set up the original approach to be
more flexible but I don't need the flexibility, and it's useful to know
that one handler has finished before the next starts. This also lets me
simplify the code significantly.