Adjust pcap loaders to avoid race conditions
These loaders were still not not race-free. The core issue is that you
shouldn't issue a kill to Golang process if Wait() has already been
called. Therefore, I need to ensure that all kills are serialized with
the result of calls to Wait(). To mitigate, the result of Wait() is send
to a channel which is consumed by a goroutine that also is the only
thread of control that will call kill for a given termshark
child-process.