Change default behavior of the Escape key
Prior to this change, the escape key set focus on the button that opens
the menu, and then also opened the menu. Now that there's a
minibuffer/cmdline, I find this Escape behavior confusing. For example,
if you're focused in the display filter and you want to run a
minibuffer command, it's instinctive to hit Escape to "leave" the focus
of the filter and then ":" to open the command-line. But Escape will
open the menu; so then you need to hit Escape to close it again before
":" for the command-line. Now, Escape will put focus on the menu button,
but not open the menu. Just hit Enter to open it. Putting focus on the
Menu button, in particular takes focus away from the filter widget, so
the command-line will still be ":" keypress away.