Workaround for incorrect colors with base16-shell and truecolor
Background: base16-shell is a project that allows base16 themes to be
used in the terminal (https://github.com/chriskempson/base16-shell). If
you are using base16-shell, the lowest colors 0-21 in the 256 color
space will be remapped to whatever colors the terminal base16 theme sets
up. If you are using a termshark theme that expresses colors in RGB
style (#7799AA), and termshark is running in a 256-color terminal, then
termshark will find the closest match for the RGB color in the 256
color-space. But termshark assumes that colors 0-21 are set up normally,
and not remapped. If the closest match is one of those colors, then the
theme won't look as expected. A workaround is to tell gowid not to use
colors 0-21 when finding the closest match.

If your environment sets COLORTERM=truecolor, then your terminal
emulator may receive truecolor-ANSI codes for colors. Termshark will
assume you're running in 24-bit mode unless told otherwise. The problem
with this, in the presence of base16, is that some terminal-emulators -
e.g. gnome-terminal - still seems to map RGB ANSI codes colors to values
0-21 in the 256-color space. I'm not sure if this is just an
implementation snafu, or if something else is going on... In any case,
termshark will fall back from 24-bit color to 256-color mode if base16
is detected, to avoid this problem. If you *really* want to use 24-bit
color and base16-shell together, you can set main.respect-colorterm=true
in termshark.toml.