Better control of a display filter error message
The pcap loader's NewFilter entrypoint is called in several places
e.g. by the stream reassembly UI. But the fact that the filter being
applied is the same as the current one is an error/warning only if the
user applies it from the filter widget. So I don't want these errors
displayed in other circumstances.