Provide a table comparator that understands IP addresses
If you have a table column that displays IP addresses, this enables the
table to sort your data in a useful way. I'll make use of this for
tshark PSML data that comprises IP addresses. IPv4 sorts before IPv6
unless the IPV6 is a V6-to-V4 mapped address, in which case they are
compared as IPv4. Valid IP sorts before invalid IP; if both are invalid,
they are compared lexicographically.