Simple utilities to track and compute averages and maximums
I use these when reading in PSML column data. The average length and
maximum length of a column are passed as configuration to the table that
displays the data. A rendered column will use no more than the column
maximum number of cells, and if that number of cells is not
available (e.g. on a narrower terminal), then the width of the column
will be determined by the column's average in proportion to the other
displayed columns' averages or rendered widths.