Connect the user-configured PSML columns to the rest of termshark
The loader will now translate the user-supplied columns and pass them as
arguments to the PSML loading process. The loader will also store
average and maximum PSML column widths to help the table to display
itself optimally. When termshark starts it will try to load valid column
identifiers from a cached version of the output of tshark -G
column-formats; if not found, it will generate it.