Move logic in termshark.go under the control of the module
I am doing this is an attempt to make go get work again with branch
targets other than master. The problem is that if you issue

go get github.com/gcla/termshark/v2/cmd/termshark@somebranch

then Go doesn't apply the replace directives from termshark's go.mod; I
think because it considers termshark.go's package main to not be part of
the module. Here's a description of the issue:

https://github.com/golang/go/issues/30354

I certainly did not appreciate this until I read it. This is an
experimental branch to see if I can go get from a branch by moving all
the logic of main() under the umbrella of termshark's go.mod.