Add the solution to Kevin's growing file issue to the FAQ
This can be handled with tail and a pipe. But this solution does still
have a downside - termshark will read the live capture and write the
packets out under ~/.cache/termshark/pcaps/ - so the file will exist
twice on disk. It would be nicer if termshark could avoid that. An idea
that I haven't fleshed out yet is to be able to just run

$ termshark -r foo.pcap

and have termshark detect if the file is likely to grow e.g. by using
something like fuser to indicate something else has it open for
write (this wouldn't work if permissions didn't line up). If it's
detected to be "live", then I can use termshark's internal mechanism to
handle foo.pcap in the same way as I'd handle the output of tshark -i if
termshark was reading from an interface.