A function to allow the current source to be reloaded directly
I haven't had an API until now because I haven't needed to reload unless
some information has changed - such as the display filter or the pcap
source. However soon the displayed columns will be configurable, so I
want to be able to redisplay. An alternative would be for termshark,
when loading the PSML, to request *every* column; save those internally,
and then just display the columns the user wants to see. I decided
against this because it would use a good deal more memory on larger
pcaps, and I don't imagine changing columns is an operation that is done
on every pcap load. So instead of making every termshark instance pay
the price by loading every column value, I just reload the pcap when the
columns change.