A model and widget for editing PSML columns
The model contains a table widget directly so that via Go's embedding,
it itself is a widget, and then when model updates are made, the table
widget can be generated directly from the model without having to have
context injected.

Each row can be moved up or down or deleted. A custom name can be
applied, which if not empty, will be used as the display name of the
column in the UI. Drop-down menus are provided for changing the current
row's meaning.