Add a menu to the packet struct view for adding a column
This mimics Wireshark's "Apply as Column" menu on the right mouse
button. The column name is computed from the @showname attribute in the
PDML. For awkward long cases like this:

showname=".... ..1. .... .... .... .... = LG bit: Locally administered address (this is NOT the factory default)"

I strip everything before and including the "=", then strip everything
after the ":".