Rethink termshark's menus
I've pushed the app awareness interface, into the menu widget -
menu.IOpener; and now I pass openTermsharkMenu to the menu widget
instead (for most uses). I need to do this because the menu can close
itself via code in the menu package, rather than just via termshark's
calls to MenuOpener.Close() e.g. if the escape key is hit. That then
leads to broken overlays. The widget itself needs to be able to close
itself in a way that is compatible with the containing application.