Extend the simple menu API to return a list of menu sites too
Now, when building a simple menu, you can pass in a map and have it
populated - the key is the widget amongst those use to build the menu,
and the value is the menu site. This site can then be used to anchor a
second menu so that it opens where the widget in the first menu was
clicked.