Generalize the conversation filter logic
This lets me use it to populate a context-aware menu that is opened when
clicking a button that is now present on the right hand side of each
entry in the packet struct view. The menu holds, as context, the display
ay filter that corresponds to that level of structure, via the PDML. The
menu lets the user (a) apply the filter as a custom column, or (b)
prepare/apply the filter as a display filter.