Bug fix - close the filter goroutines when the columns dialog closes
The columns config dialog shows a display filter validation widget for
each custom column displayed. This launches two goroutines per
filter. These should be cleaned up when the dialog closes, so as not to
leak resources. But if the dialog is closed via the escape key, for
example, the Close button logic doesn't run. The solution is to set a
callback via OnOpenClose() and shut down the filter goroutines there.