More closely imitate Wireshark's contextual filter menus
If you right click on part of the packet structure in Wireshark, you can
apply a filter based on that structure. For example, applying a filter
based on UDP Source Port might result in

udp.srcport == 63801

But my recent contextual menu would apply only

udp.srcport

Now I include the value that should match the filter, when it's
available.