Send termshark's current pcap with magic wormhole
This commit implements a new console command, "wormhole", that allows
you to transfer termshark's open pcap using magic wormhole -
https://github.com/magic-wormhole/magic-wormhole. Magic wormhole is a
secure and convenient way of getting files from one machine to
another. The sender generates a wormhole code e.g.

  $ wormhole send foo
  Sending 1.9 kB file named 'foo'
  Wormhole code is: 9-phonetic-goldfish
  On the other computer, please run:

  wormhole receive 9-phonetic-goldfish

and then the receiver references the code:

  $ wormhole receive 9-phonetic-goldfish
  Receiving file (1.9 kB) into: foo
  ok? (y/N): y
  ...
  Received file written to foo

When you run termshark's wormhole command, termshark generates a
wormhole code and displays it in a dialog. You can then use magic
wormhole locally to transfer the pcap to your desktop. The purpose of
this is to make it easy to get a pcap into Wireshark. Termshark can be
fine for a quick analysis, but Wireshark is much more powerful :-)

To make this process even easier, you can pair this new command with the
tmux-wormhole tmux plugin. You can find it here:
https://github.com/gcla/tmux-wormhole. This can help if your tmux is
local, and within that tmux you are sshed to a remote machine and using
termshark there. As long as the remote machine has internet access, you
should be able to magic-wormhole the pcap. Once you see termshark's
wormhole code on the screen, hit the plugin hotkey - C-b w by default -
and the plugin will detect the code and let you download and open the
pcap. If your desktop association for pcap files is Wireshark, then it
should open in Wireshark. Here's a demo of it in action:

https://drive.google.com/file/d/16qCXyjWS8smzjOeJLGZiplib3frgK4F5/view?usp=sharing

This implementation uses psanford/wormhole-william, a Go implementation
of the magic wormhole protocol.