First attempt to fix loader bug that means termshark doesn't quit
This is an attempt to solve issue #121. I think my mistake is as
follows:

- I start the goroutine that runs the pcap process
- after starting the process, I send its pid via a channel to
  another goroutine that controls process life-cycles
- however that goroutine has closed down because the user issued
  a cancel

The process lifetime goroutine should not shut down if it is guaranteed
the pcap running goroutine is going to send its pid, because then
nothing will listen for it. I've adjusted the logic so now that
goroutine will definitely wait for the pid - unless it receives a pid==0
which indicates there was a problem starting the process.