Bug fix - when main.column-format is not set and a column is added
The packet structure view has a new button at each level "[=]" that can
be used to add the field as a custom column. If the user has no custom
columns set, then termshark will use a default set. But if the user
added a custom column, I didn't append to a copy of the default set if
no custom columns were configured. This meant that on the addition of a
custom column, termshark would show only the custom column, and none of
the original columns.