Bug fix for -w flag usage
If termshark is invoked like this:

$ termshark -i eth0 -w foo.pcap

then termshark will save the captured packets to foo.pcap. It checks to
make sure that foo.pcap is not something that tshark will complain
about, but I mistakenly did not bypass that check if foo.pcap does not
exist, which will be the typical case.