Various improvements to the progress bar
With termshark v2.2, progress could be "jumpy" sometimes, flickering on
and off again. Here is the intended behavior:

- If PSML is being loaded, use the progress through the pcap file as the
  progress bar percentage
- If PDML is being loaded and is "visible" i.e. the UI is waiting for
  packets in order to display them, use the ratio of the current length
  of the loaded 1000 packet batch to the packet number needed to
  populate the display as the progress bar percentage

If the /proc support to determine progress through a file descriptor is
missing on the target OS, a spinner is displayed instead.

Note that if an interface, fifo or pipe is in use, the progress bar will
be a spinner until/unless the read is interrupted; then will switch to
progress until the PDML has been processed far enough in order to
satisfy the UI's need for the packet struct and hex views.