Fix bug that prevented termshark termination on SIGTERM
If the columns configuration widget is open when termshark receives a
termination signal, termshark's UI will shut down but then the process
blocks. This is because I programmatically insist all goroutines have
stopped before I terminate, but the columns widget needs to be closed to
stop its goroutines. On consideration, I'll remove the goroutine check
for official releases; but the fix for this bug is, similar to the
wormhole widget, to close the dialog on receipt of a termination signal.