Speed up the display filter
I was revalidating the display filter every time the cursor was moved
within the filter widget, but this is unnecessary - it only needs to be
revalidated on change. This speeds up the experience of using the
display filter, which is also deployed in the custom columns dialog box
now too.