An 8-color theme
This is in support of issue #114. I have a patch to tcell that turns
8-color terminals into 16-color terminals using the foreground-bright
trick: https://en.wikipedia.org/wiki/ANSI_escape_code; BUT I wasn't
using it. So when termshark runs under TERM settings like xterm or
screen, it looks for an 8-color theme (not 16), but one isn't provided
by default. This was an oversight! I hope this theme is reasonable for
both regular mode and dark mode.