Help the user upgrade to a more colorful TERM setting
If the user runs in a terminal emulator with TERM=xterm or TERM=screen,
then termshark will only use 8-colors. This is fine but the experience
is better with 256 colors available (or more) - for example, for packet
coloring. If the user knows to change TERM to xterm-256color, that's
also fine, but not all termshark users will know to do this. This change
adds a workflow to help the user run in a higher-color mode, if
possible. On startup, termshark will check to see if (a) only 8 colors
are in use and (b) if a 256-color version of the TERM variable is
available. If so, termshark opens a dialog to suggest switching. If the
user hits "No, don't ask", then termshark sets main.disable-term-helper
in termshark.toml and won't ask again. If the user hits yes, termshark
will restart using the 256-color TERM. After 3 seconds, termshark will
open a dialog to ask the user to confirm that everything looks ok - the
default is no, and if the user hits that, or if 10 seconds elapse
without input, termshark reverts back to the original setting. If the
user hits yes, termshark will ask if the user wants to save that TERM as
the default to use; if yes, main.term will be set in termshark.toml.