When advising on a TERM upgrade, use tcell's database too
Tcell comes with a built-in database of terminal types. It can also use
infocmp to construct one on the target machine. Termshark will advise a
user to use a more colorful TERM variable if the -256color definition is
available. This change has termshark check tcell's DB first, falling
back to infocmp if necessary.