Add a new config suppress-tshark-errors to hide parsing problems
I saw this PDML error on macos with tshark 3.4.4:

<field name="irc.response" showname="Response: :Samsung.US.SPUNet.org NOTICE Auth :Welcome to \002SPUNet\002!" size="56" pos="54" show=":Samsung.US.SPUNet.org NOTICE Auth :Welcome to SPUNet!" value="3a53616d73756e672e55532e5350554e65742e6f7267204e4f544943452041757468203a57656c636f6d6520746f20025350554e65740221">

The Go XML parser complains about the 0x02 bytes surrounding SPUNet. If
this happens, by default, termshark will display an error dialog. On
this particular PCAP, the error was triggered by many packets, so
navigating around the PCAP popped up the error repeatedly. This config
option allows the user to suppress these errors. The setting is stored
in termshark.toml as main.suppress-tshark-errors.