When capinfo is opened, put focus on the Close button
I expect the user will typically want to look then close, so this is
more streamlined. However, if the user navigates up so that they can
enter copy-mode and copy the capinfos data, I now use a "slim" frame
widget around the capinfos text to indicate that there is something
there that can be selected and copied.