Update to latest gowid and to tcell/v2
Thanks to @hinshun, gowid is now built against tcell/v2. This is
interesting to me because it more easily enables UI testing of termshark
via:

https://github.com/gdamore/tcell#testability

That will allow me to codify - some of - the current behavior of
termshark and so have more confidence that future changes and additions
don't break any of the current features.

This also lets me remove one of my replace directives, so brings me one
step closer to being able to give a one-line go install instruction for
installing termshark!