Bug fix to restore preservation and display of recently-used filters
I inadvertently broke this - when you place focus in the display filter
widget, with an empty field, the drop-down should be populated with the
list of most-recently-used filters, like Wireshark. Instead, I wasn't
saving these filters, so the list to display was always empty - in that
case, termshark falls back to show filter tokens whose prefix matches
the current field value (which is empty).