Make the hex view's cursor position more visible
Usually, inside the hex view of a packet, the user will see the cursor,
then a colored layer representing the current "field", then another
colored layer representing the next level up in the packet
structure. Termshark's themes make the cursor easy to spot within the
boundaries of the current packet field in the UI. But sometimes the PDML
returned by tshark does not cover the entirety of the packet's
bytes. That means that termshark's cursor can be moved to a spot outside
the bounds of the last highlighted layer, and it's not always to easy to
spot. This is a small tweak to make the byte under the cursor show up
bold.