A new tree iterator that disregards expanded and contracted nodes
Termshark currently uses a tree iterator for moving around the packet
structure view widget. The iterator's Next() API takes into account
whether or not the user has expanded the node in the UI. If the user
hasn't, then the iterator moves to the sibling node. If the user has
expanded the node, the iterator descends a level to the first child.

For termshark's upcoming search feature, I need to be able to iterate
over the packet struct data without consideration for whether the nodes
are expanded or contracted in the UI - because the search feature looks
throughout the whole packet. This commit provides a new iterator which
can be easily created from an existing tree model. A new tree model type
called ExpandedModel has been added that is equal to
tree.Model. Therefore, if you have a tree.Model, you can simply
type-cast it to tree.ExpandedModel. From there, the iterators returned
will be tree.ExpandedIterators.