Don't activate the termshark tcell screen right away
This was the behavior as of v2.3.0. The reason is that if the user
starts termshark on an interface, often the host machine is not set up
to allow this without superuser privileges. If termshark is configured
to e.g. run tshark under sudo, then there may be a prompt on the
terminal for a password. If the termshark UI starts right away, this
prompt will be obscured and the user won't have any indication as to why
packets aren't being shown. So termshark doesn't activate the screen
until it detects packets, and it displays

(The termshark UI will start when packets are detected on eth0...)

in the meantime.

I broke this behavior with a recent gowid change, causing termshark's
screen to be activated immediately on startup.