Add a new termshark filter state
Prior to this change, termshark's display filter widget could be either
Valid, Invalid or Indeterminate (while tshark runs to validate the
filter) But when the filter is empty, there should be a special Empty
state - it's neither Valid not Invalid, but can't be applied. Wireshark
represents this with a white background. For termshark's upcoming search
feature, I am going to re-use the filter widget and will have it
validate other syntaxes where I would also like to distinguish Empty
from Valid/Invalid. So this change adds an Empty state, and the UI
plumbing required to display it. By default, the empty filter is shown
with a cyan background - I think this looks better in a default
terminal. There is a new filter-empty theme field that can be tweaked if
you don't like the defaults.