Try harder to make it possible to go install termshark
This detail of Go's replace directive means that I cannot go install
termshark:

- No module is considered the "main" module. If the module containing
packages named on the command line has a go.mod file, it must not
contain directives (replace and exclude) that would cause it to be
interpreted differently than if it were the main module. The module must
not require a higher version of itself.

https://pkg.go.dev/cmd/go#hdr-The_go_mod_file

Termshark's go.mod has a replace directive for pkg/term to a fork I
posted, gcla/term, with some stub functions needed to compile on
openbsd (they aren't called by termshark, but the compiler needs
them). These changes wouldn't be suitable for upstream, since they don't
add the needed functionality - so I haven't tried to get them
merged. But the presence of this replace directive means I cannot now go
install termshark directly, which is not ideal. I'd like it to be as
easy as possible for users to grab termshark and compile it.

I don't like the idea of hosting gcla/term because it implies an
authorship that isn't true - none of pkg/term is mine. But I'm not sure
if there is a better solution that lets me remove the replace directive.