Silly error in the edit widget led to losing cursor control
This is reproduced from the gowid commit ID
4d04cba8013a7ba78840387f8a45f2ebe05c7809

The UserInput function for the edit widget was returning false even in
cases where it was handling the input. If e.g. edit was embedded in a
column widget with selectable widgets either side, a left/right cursor
press inside widget would move the widget's cursor, but also bubble up
to the column widget and cause it to change the focus widget from edit
to one of its neighbors.

In the context of termshark, this meant the various filter widgets were
losing focus when the user moved the cursor around the user-entered
text.